scripts/requirements.txt
pillow>=12.0
playwright>=1.40
